requirements.txt
numpy==1.26.4
pandas==2.1.4
torch==2.2.2
tqdm==4.65.0
scikit-learn==1.3.0
py-ecg-detectors
scipy==1.11.4
similaritymeasures==0.7.0
frechetdist
adapt
keras==2.11.0
tensorflow==2.12.0
keras_resnet
focal-loss==0.0.7
